move profile/index.js to profile/profile-header.js
